pyproject.toml
[project]
name = "rag-mini"
version = "0.1.0"
description = "A tiny RAG app (OpenAI + FAISS) for learning proper project setup."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.0.0",
  "faiss-cpu",
  "numpy",
  "typer",
  "python-dotenv",
]

[project.scripts]
rag-mini = "rag_mini.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"